main: eliminate the --git-path option
1d431b6db12fc2f99d3499708e9a37cc09e606ad taught git-cola to
automatically find Git/bin when running on Windows, so there's no need
for a command-line flag to specify the path to Git anymore.

Add support for ~/.config/git-cola/git-bindir so that users who have Git
installed in a custom location, but do not have Git/bin/ in their $PATH,
can write its location into this file and have it automatically added to
their $PATH by git-cola.

Made-possible-by: #421
Signed-off-by: David Aguilar <davvid@gmail.com>