cmds: ensure that GIT_EDITOR is set for all rebase commands
The Rebase->Continue, Rebase->Edit Todo, and Rebase->Skip actions were
not setting the GIT_EDITOR variable, which can cause git to use the
wrong editor during the rebase.

The previous commit ensured that GitXBaseContext always provides
GIT_EDITOR in the environment, which fixed the "Edit Todo" command.  Not
providing GIT_EDITOR can cause Git to use the default `core.editor`,
which is typically intended for terminal sessions.

Use the GitXBaseContext in the remaining commands so that the "Continue"
and "Skip" actions use the correct editor while rebasing.

Closes #445
Reported-by: Matthias Mailänder <matthias@mailaender.name>
Signed-off-by: David Aguilar <davvid@gmail.com>