widgets/diff: rewrite syntax highlighter
The previous DiffSyntaxHighlighter worked adequately, but because it was
stateless, it needed a lot of complicated regular expressions to try to
match all the possible diff header lines.  It could also be thrown off,
if for example a line was removed starting with '-- ', the resulting
diff line would start with '--- ', causing the syntax highlighter to
interpret it as a header line instead of a removal.  To address these
shortcomings, completely rewrite the syntax highlighter to be stateful.
This greatly simplifies the logic and drastically reduces the number of
regular expressions that need to be matched.  Because
DiffSyntaxHighlighter was the only user of GenericSyntaxHighlighter, and
GenericSyntaxHighlighter does not fit the stateful approach, eliminate
GenericSyntaxHighlighter (moving a few necessary bits to
DiffSyntaxHighlighter) and move DiffSyntaxHighlighter from qtutils to
widgets.diff, where it is actually used.

Closes #462
Reported-by: Andreas Sommer <andreas.sommer87@googlemail.com>
Signed-off-by: Daniel Harding <dharding@living180.net>