status: improve shortcuts for the Delete action
We were previously assigning "Ctrl+Backspace" to "Delete".  This was
later adjusted to use "Alt" instead of "Ctrl", but that was a step
backwards.

Originally, when `send2trash` was installed, we would assign the
"Ctrl+Backspace" hotkey to "MoveToTrash", and "Ctrl+Shift+Backspace" to
"Delete".

This is problematic for several reasons.  First off, we should not
clobber the prevalent "Ctrl+Backspace" hotkey, as it is used to remove
the previous word in text edits.

Secondly, we don't need a shortcut for *both* "MoveToTrash" and "Delete"
when the `send2trash` module is installed.  Using a single shortcut is
sufficient.

Adjust the logic so that "Ctrl+Shift+Backspace" is used as the hotkey.
When `send2trash` is installed, only the "MoveToTrash" command will be
bound, otherwise the "Delete" command is bound to the hotkey.

Closes #453
Helped-by: Kirit Sælensminde <k@kirit.com>
Helped-by: Taylor Braun-Jones <taylor@braun-jones.org>
Signed-off-by: David Aguilar <davvid@gmail.com>