widgets.main: do not add actions at the top-level widget scope
Remove the redundant adding of actions to the main widget.  This is
needed to avoid ambiguous shortcut overloads with the Status widget.

This real benefit of this is that this avoids clobbering the
Ctrl+Backspace hotkey, which frees us to use any hotkeys we want in any
other widget.

This is especially important because the diff editor really should have
vim hotkeys.  Fix it once and for all.

Closes #453
Signed-off-by: David Aguilar <davvid@gmail.com>