diff: add Alt-J and Alt-K Previous/Next File shortcuts
The j/k shortcuts no longer bubble up to the status widget from the diff
editor now that the diff editor understands vim navigation keys.

Add Alt-J and Alt-K shortcuts to the diff editor.  These
shortcuts emit a signal that is routed through the main widget and back
to the status view, which then performs move_up() and move_down().

Signed-off-by: David Aguilar <davvid@gmail.com>