prefs: better history browser defaults on Windows
Use "git cola dag" as the default history browser on Windows.  This
fixes launching the history browser on Windows for new users, and avoids
trying to launch `gitk`, which fails on Windows because it's a shell
script.

We now configure things so that the defaults work correctly on Windows.
Unix setups are unchanged.

Mention how to setup things for `gitk` in the README.

Closes #429
Reported-by: deniz1a <deniz1a@deniz1a>
Helped-by: geotavros <geotavros@gmail.com>
Signed-off-by: David Aguilar <davvid@gmail.com>