widgets: persist the clipboard when shutting down
The copy/paste clipboard does not persist after git-cola exits.

This happens because X11 does not store a copy of the clipboard
contents, it only stores a reference.  The reference is destroyed when
git-cola exits.

Send a clipboard event to trigger a copy of the cilpboard data out of
the application.

C.f. https://stackoverflow.com/questions/2007103/how-can-i-disable-clear-of-clipboard-on-exit-of-pyqt4-application

Closes #484
Reported-by: Josh Taylor <joshuataylorx@gmail.com>
Helped-by: Ｖ字龍(Vdragon) <Vdragon.Taiwan@gmail.com>
Signed-off-by: David Aguilar <davvid@gmail.com>