browse: do not auto-delete tasks
QThreadPool automatically deletes tasks when setAutoDelete is True and
QThreadPool::run() is called.

This causes crashes on some versions of PyQt because Python and Qt are
both trying to free the same task.

Tell Qt to not delete the tasks so that Python's garbage collector can
take care of it.

Signed-off-by: David Aguilar <davvid@gmail.com>