app: add the correct path to git-cola's bin/ directory
We were using `__file__` to get the path to the launching script, but
`__file__` points to git-cola's module directory.

The original code was refactored out of `bin/git-cola`, where it was
originally valid.  Adjust the incantation to use `sys.argv[0]` instead,
which restores the original behavior.

This also fixes problems when using a `PATHEXT` that includes `.PY`, as
that would cause git-cola to find its `git.py` module rather than
`git.exe`.

Closes #492
Reported-by: Bert Jacobs <albert.devs@gmail.com>
Signed-off-by: David Aguilar <davvid@gmail.com>