status: use Italics instead of Bold-with-background for headers
The midlight depends on the theme, and is not always readable.  Simplify
the headers by using an Italic font.

This should keep things both readable and distinguishable.  Bold was
dropped to improve readability.  I would have used SmallCaps, but I
recall some Windows users having readability problems with SmallCaps, so
keep it simple.

Signed-off-by: David Aguilar <davvid@gmail.com>