bookmarks: add a "Default Repository" feature
Allow setting the "Default Repository" by right-clicking in the
"Favorites" or "Bookmarks" widgets.

The "Default Repository" is used when the working directory does not
contain a Git repository.  If it is not set, or if it is set to a bogus
value, then `git cola` will still prompt for a repository.

Closes #513
Suggested-by: Zeioth <zeioth@hotmail.com>
Signed-off-by: David Aguilar <davvid@gmail.com>