git-gui--askpass: generalize the window title
git-gui--askpass is not only used for SSH authentication, but also for
HTTPS. In that context it is confusing to have a window title of
"OpenSSH". So generalize the title so that it also says which parent
process, i.e. Git, requires authentication.

Signed-off-by: Sebastian Schuberth <sschuberth@gmail.com>
Reference: http://article.gmane.org/gmane.comp.version-control.git/285200
Signed-off-by: David Aguilar <davvid@gmail.com>