models: consolidate behavior in set_worktree()
We must reset the gitcfg, chdir, and set attributes whenever the current
repo changes.  Consolidate this behavior into set_worktree so that the
other callers do not need to worry about these details.

This fixes an oversight where launching from a desktop icon would cause
cola to not see the repo's config.  Ensuring that both chdir() and
reset() are called ensures that the config is properly read.

Closes #548
Reported-by: Ｖ字龍 <Vdragon.Taiwan@gmail.com>
Signed-off-by: David Aguilar <davvid@gmail.com>