setup.py: allow installing from and to utf-8 encoded paths
Python's stdlib is buggy as all hell and it does not properly handle
unicode.  The *only* crazy thing that does work is to force the encoding
to utf-8 on python2 to fix their broken code.

Allow installing to/from utf-8 encoded paths by making python2 have the
same utf-8 default as python3.

Related-to: #551
Signed-off-by: David Aguilar <davvid@gmail.com>