app: cleanup temporary directories immediately after use
We were relying on re-using a temp directory so that we could
easily cleanup that directory when exiting.  This is problematic
when git-cola is exited via a signal handler, as it does not
have a chance to cleanup.

Instead of relying on a deferred cleanup, make the individual
commands cleanup after themselves.  This removes the need to
cleanup as the application is shutting down.

Closes #566
Reported-by: Jordon Bedwell <jordon@envygeeks.io>
Signed-off-by: David Aguilar <davvid@gmail.com>