fsmonitor: remove the refs_only parameter
It turns out that using the refs_only parameter for the start() method
of the _Monitor class was inadequate for preventing unwanted refreshes
in the DAG window.  It worked when running git-dag directly, but when
the DAG window as launched from within git-cola, the undesired refreshes
were still present (because in that case the start() method is called
with refs_only=False).

Related-to: #516
Signed-off-by: Daniel Harding <dharding@living180.net>