fsmonitor: avoid refreshing for ignored files
Changes to files ignored by git won't cause display changes in git-cola,
but they were still causing the file system change monitoring to refresh
the display.  This commit avoids this behavior by using git check-ignore
with if git version 1.8.5 or later is present.  The logic in the file
system change monitor is enhanced so that it no longer triggers a
refresh on any filesystem change.  Now, if the change is for a file, it
adds the file path to a set of paths that gets checked in the
_BaseThread.notify() method.  This method passes the paths to git
check-ignore, and only triggers a refresh if at least one of the paths
is for a non-ignored file.

Fixes: #517
Related-to: #516

Reported-by: Fabio Coatti <fabio.coatti@gmail.com>
Reported-by: Stanisław Halik <sthalik@misaki.pl>
Signed-off-by: Daniel Harding <dharding@living180.net>