Don’t make userDataWrapper a subclass of QObject
PySide *can* deal with Python objects, but issues occur when those objects implement __getitem__. For instance, if __getitem__ raises exceptions, this causes PySide to segfault. If __getitem__ returns valid values, it can cause PySide to hang. So we just need to wrap it in a simple Python class to ensure that __getitem__ isn’t defined.