cola: support PyQt5, PyQt4 and Pyside via QtPy
Use the vendored QtPy library to add support for PyQt5 and Pyside.

git-cola is now compatible across PyQt4, PyQt5, and Pyside.
Future QtPy work will add Pyside2 compatibility.

Qt5 contains known bugs that prevent us from recommending it as a stable
target, but this makes it available for users that want to start testing
the waters.

Existing users are not required to install any additional libraries
because we contain a vendored copy of QtPy.  Users that want to use
PyQt5 must install it separately.

Closes #232
Signed-off-by: David Aguilar <davvid@gmail.com>