core: handle utf-8 paths when LC_ALL=C
Python requires that we encode the current directory when LC_ALL
is set to the "C" locale on python2.

Furthermore, we cannot rely on `os.getcwdu()` since it can be tricked by
the user's environment, and relies on heuristics that are outside of our
control.

Go back to the original behavior before
de13a639e59593c8033d91afb9a240d4db0dc912 switched over to
`os.getcwdu()`.

Closes #581
Reported-by: Ｖ字龍 <Vdragon.Taiwan@gmail.com>
Signed-off-by: David Aguilar <davvid@gmail.com>