cmds: fix a subtle single-char typo when switching repos fails
When switching repositories fails we fallback to the original
path.  We called `gitcwd()` when we meant to call `getcwd()` to
get the path, but the call to the mistyped name did not fail as
one might expect.  The Git object forwards all unknown method
calls to `git`, so it returned the tuple result from running
`git getcwd`.

Fix the typo so that the `getcwd()` method is called.

Closes #606
Reported-by: @MikHulk via github.com
Signed-off-by: David Aguilar <davvid@gmail.com>