git: replace all references to "sha1" with "oid"
Git is internally renaming all "sha1" and similarly-named data
structures to "struct oid" (object ID).

Adjust the code base so that it refers to object IDs (oid) instead of
SHA1-s for consistency with Git.

We still keep "SHA-1" around in the user-facing strings since
translations already exist, and it's still widely used.

Signed-off-by: David Aguilar <davvid@gmail.com>