Revert "app: Make homebrew work by default"
This change was introduced to fix Homebrew/legacy-homebrew#7462, which
describes an installation issue that is specific to Homebrew and not to
macOS in general. If Homebrew really needs this workaround, this change
should be applied as a local patch instead of forcing /usr/local to be
added to sys.path for everyone.

MacPorts has to remove this addition to sys.path, see
macports/macports-ports@e215228a4a18d51ae099e63ad1b416757af64150

This reverts commit 9674055ba59182c488e32d225331dbcda6ba3afd.

Signed-off-by: Rainer Müller <raimue@macports.org>