po: zh-TW: Replace "repository" translation to "版控庫"
Previously we used "版本倉庫" as the translation of "repository" which is
literally "repository of Versions", however according to "Repository
(version control) - Wikipedia" repository in VCS is "an on-disk data
structure which stores metadata for a set of files and/or directory
structure." which is far more complex than simply "repository of
versions".

This patch replace the translation with "版控庫(repository for version
controlling)" which should be much appropriate.

Refer-to: https://en.wikipedia.org/wiki/Repository_(version_control)
Signed-off-by: Ｖ字龍 <Vdragon.Taiwan@gmail.com>