difftool: move low-level command logic into the cmds module
We want to use the "Edit" command from within the difftool widgets, but
doing so would cause a circular dependency because the `cmds` module
uses the `difftool` module.

Refactor the code so that the low-level command logic lives in the
`cmds` module.  Sever the dependency from `cmds` on the `difftool`
module so that we can `import cmds` cleanly from the `difftool` module.

Signed-off-by: David Aguilar <davvid@gmail.com>