widgets: revamp text widgets
Refactor the text widgets to eliminate the redundant "View" textedit
classes.  This was driven by waning to use QPlainTextEdit as our base
widget.

Rework how the hinted text is handled on PyQt5.  When
setPlaceholderText() is available we skip a lot of the complicated logic
and handle error coloring only for the hinted widgets.

Use the new widgets to give the grep view a line numbered view.  A
similar technique can later be used to add line numbers to the diff
editor.

Signed-off-by: David Aguilar <davvid@gmail.com>