diff: trim the selection to make it easier to select lines
If the cursor is sitting at the beginning of a line, and nothing from
that line is selected, then the line should not be considered selected.

The end result is that we can use "Shift+DownArrow" in the diff editor
from the left-most column and it will only stage the fully selected line
above the cursor, which is likely what the user intended.

Signed-off-by: David Aguilar <davvid@gmail.com>