browse: vastly speed up File Browser performance
Teach the File Browser to incrementally load data rather than trying to
build items for every file in the repository up front.

This alone vastly speeds up the File Browser since we will only build
items for the root of the repository.

Use the canFetchMore() and related APIs so that Qt will progressively
load data.  Make it so that expanding an item is now responsible for
populating the directory with children.

Additionally, add a new "turbo" mode that is enabled by setting
`cola.turbo` to true.  This mode will skip loading background Git commit
data in the File Browser.  Specifically, the Message, Author, Age, and
Date fields will not contain any data in Turbo mode.

Closes #499
Reported-by: Justin Lecher <jlec@gentoo.org>
Signed-off-by: David Aguilar <davvid@gmail.com>