browse: simplify item ownership to eliminate C++ object deletion
We were allowing signals to be delivered to dead C++ objects as a
side-effect of using a global GitRepoEntryStore to keep track of
entries.

The global entries contained signal/slot connections that connected
directly to the actual RepoItem to update, which created a subtle
problem as the items are allowed to be deleted at any moment.

Rework the notification scheme so that the background thread posts
events to the RepoTreeView rather than to a global entry.  Make the tree
view in charge of looking up entries and updating them.

Pass the path in the message and update the entire row in one event
rather than sending separate events for each column.  This reduces the
event overhead, thus improving performance further.

Signed-off-by: David Aguilar <davvid@gmail.com>