dag: revert hotfix of undefined frontier
The propagate_frontier assumes that frontier column is always defined. KeyError
exception is an indicator of an error in another part of the grid layout
algorithm. So, do not except KeyError to get the error as soon as possible.

The hotfix may be applied again for temporary error elimination.

At the other hand, defining frontier in this place too, complicates the
algorithm logic.

Signed-off-by: Efimov Vasily <real@ispras.ru>