tests: remove the slowest test
EINTR is hard to test, and the only flaky method we were using requires
the test to  sleep() in the test before executing the unix `sleep`
command for an entire second.

signal.alarm(int) has 1-second resolution, which is why the external
command sleeps for 1 second.  The test sleeps for 0.1 seconds to allow
time for the alarm to get delivered, which totals a minimum of 1.1
seconds waiting for the test to complete.  Ditch the slow test.

Signed-off-by: David Aguilar <davvid@gmail.com>