gitcfg: make get_all() return results sorted in priority order
Allow users to set cola.icontheme in their local repos.  Previously,
get_all() would return results ordered according to how git saw them,
which causes a problem because git scans the files in a way where the
highest priority items are listed after the lower priority values.

We still want to keep git's relative ordering within a specific config
file, so teach get_all() to group the results according to their source,
and then apply the results in the "reverse" order at the group level.
This effectively makes it so that the local configuration's values are
listed before global and system values.

Signed-off-by: David Aguilar <davvid@gmail.com>