completion: avoid Runtime Error: C++ object has been deleted
Catch runtime errors for the rare case when completer threads take too
long to execute and outlive the widget that created them.

Attempt to dispose of the thread when the widgets dispose, but we still
have to catch the exception because there is no guarantee that dispose()
will ever be called.

Signed-off-by: David Aguilar <davvid@gmail.com>