cmds: update the gitcfg during the Refresh command
Now that we avoid updating the gitcfg object in gitcfg's get(), we need
a convenient place to force the config object to run update().

We only need to run update() to detect configuration changes that
happened behind our back, either by the user running `git config`
commands, or, manually editing .gitconfig.

We already call update() internally in the gitcfg module when setting
values, so this is purely here so that users can tell us that want cola
to refresh.

Signed-off-by: David Aguilar <davvid@gmail.com>