tag: fix the unsigned tag fallback
The "Create Tag" dialog has a convenience feature where it will
offer the choice of creating an unsigned tag when "Sign Tag" is
selected, but no message is provided.

This feature was not correctly managing the `git tag` command-line
flags, causing it to fail.  Ensure that the `--sign` flag is properly
cleared when falling back to an unsigned tag to prevent git from
spawning an editor for the tag message.

Closes #696
Reported-by: Ｖ字龍 <Vdragon.Taiwan@gmail.com>
Signed-off-by: David Aguilar <davvid@gmail.com>