travis: use fedoraproject.org instead of sourceforge.net
sourceforge.net is down, and there is no ETA for when downloads will
become available again.

Switch to fedoraproject.og for now, but it might be worth exploring a
better long-term solution.

If we dropped all Python2 builds we could use "pip" to install PyQt5
natively, but we're not ready for that yet.

Signed-off-by: David Aguilar <davvid@gmail.com>