main: fix ambiguous shortcut overloads
Use Qt.WidgetWithChildrenContext as the shortcut context for the actions
added to the edit menu.

We were using the built-in QMenu.addAction(), which defaults to
Qt.Window.  This causes ambiguous shortcut overloads since they'll clash
with the shortcuts in the various dock widgets in the same window.

Use the qtutils.add_action() helper, which, in addition to improving
code consistency, also fixes the issue because it sets the shortcut
context to Qt.WidgetWidthChildren.

Closes #767
Reported-by: @plsamuel on github.com
Signed-off-by: David Aguilar <davvid@gmail.com>