main: support Alt-key navigation for the File and Edit menus
Allow Alt+F and Alt+E to activate the File and Edit menus.

The Qt convention is that the labels should use the "&" prefix before
the character that should be used as the hotkey.  In order to
accomodate this, the localizable string was changed to "&File" and the
existing translation were updated to use it wherever possible.

Closes #759
Suggested-by: Ｖ字龍 <Vdragon.Taiwan@gmail.com>
Signed-off-by: David Aguilar <davvid@gmail.com>