git: add support for worktree "commondir" files
This fixes getting and setting local configuration within worktrees
created by "git worktree".

Teach the gitfile reader to recognize "commondir" files.  These files
are present in the worktree-specific "git" directory that is pointed to
by a worktree's ".git" file with the line:

        gitdir: /path/to/.git/worktree/repo

Contained within .git/worktree/repo is a file called "commondir" which
can contain paths that look like "../..", pointing back up to the common
".git" directory.

In this scenario, the common ".git" directory is where the repo-local
"config" file is stored.  This is critical for getting the right
configuration bits for git-worktree repos.

Teach find_directory() about the "commondir" feature so that git_path()
can use the commondir when asked to find paths relative to the git dir.

Signed-off-by: David Aguilar <davvid@gmail.com>