status: handle Ctrl+S for "Staged", "Modified", and "Untracked" headers
Change the behavior of the Ctrl+S hotkey so that it has an effect when a
header item is selected in the status tree.

Stage all modified files when the "Modified" header is selected.  Stage
all untracked files when the "Untracked" header is selected.  Lastly,
unstage all files when the "Staged" header is selected.

Closes #772
Reported-by: @birgersp via github.com
Signed-off-by: David Aguilar <davvid@gmail.com>