darwin: remove the hacky "symlink to python" hack
We were symlinking to Python through a symlink called "git-cola" to
trick macOS into displaying "git-cola" in the titlebar.

This hack causes problems for macOS users on homebrew, which is our main
macOS deployment target, so remove these things in favor of a simpler
installation.  There's a chance that the titlebar might now say "Python"
instead of "git-cola", but that's an acceptable compromise.

Closes #636
Noticed-by: @cfrancois7 on github.com
Signed-off-by: David Aguilar <davvid@gmail.com>