Improve German translation
This improves the German translation by bringing it more in-line with
the wording of the official Git Book and removes some ambiguous
translations. It also replaces most occurrences of Git technical terms
with their English original as the German Git Book does. It should be
more clear now what each action really does.

Further work will follow while I'm using the software but the current
stage should be suitable for merge.

The po file was edited with poedit2 so it may result in more diff lines
than a hand-crafted edit. But at least the file now follows poedit2
format.

Signed-off-by: Kai Krakow <kai@kaishome.de>