core: ensure that unencoded strings are encoded
Using type(string) is incompatible with the UStr work which
was done in 0a2a011fa3eb3464644785b878f34aa654504218 towards
the solution for #751.  Those changes introduced the notion of
a unicode string that is aware of its original encoding, UStr,
and thus we should never use type(string) to check its type.

Use isinstance(obj, ustr) so that UStr instances are encoded
when going through `encode()`.

Closes #807
Reported-by: 林博仁 <Buo.Ren.Lin@gmail.com>
Signed-off-by: David Aguilar <davvid@gmail.com>