xbase: save empty files when GIT_XBASE_CANCEL_ACTION=save
Cancelling a rebase from the `git xbase` editor typically leaves the
instruction sheet file unchanged and exits with a non-zero exit code.

We want to have a mode where cancelling will save an empty file, which
will tell Git that there's nothing to do.

Teach git-xbase to save an empty file when the GIT_XBASE_CANCEL_ACTION
environment variable is set to "save".

Related-to: #826
Signed-off-by: David Aguilar <davvid@gmail.com>