status: Scroll current item into view when selection changes
When moving selection up/down with keyboard in a long file list the
selected item goes out of view and there's no way to tell what file the
current diff corresponds to.
This change ensures that latest selected item is always visible.
The horizontal scrolling is preserved as it was implemented before.

See also issue #494 and pull request #178 for related discussion.

Signed-off-by: Petr Gladkikh <petrglad@gmail.com>