diff: fix changing diff viewer options
In commit 649de5bb248c3a33e6778c44f75ab4457e13508c the diff viewer was
refactored to be able to display images diffs.  As a result of that
change the "parent" passed to the Options object is now a Viewer rather
than a DiffEditor.  In the Options.update_options() method, it calls the
update_options() method of the "parent" provided when it was
constructed.  However, unlike DiffEditor, Viewer did not have an
update_options() method, which would cause an exception when
Options.update_options() was executed.

To address this add an update_options() method to Viewer which delegates
to the update_options() method of the contained DiffEditor.

Fixes: #833

Signed-off-by: Daniel Harding <dharding@living180.net>