main: display an error dialog when git cannot be executed
When git is not installed we were printing an error message to stderr
and exiting early.  Defer this error handling until after the GUI is
initialized so that we can display an error message graphically.

stderr goes to a log file when running under the Windows installer so
it's best to display the error in both places.

Closes #830
Reported-by: 林博仁 <Buo.Ren.Lin@gmail.com>
Signed-off-by: David Aguilar <davvid@gmail.com>