logo: Drop unnecessary attributes
Eliminate attributes that are only local machine relevant, privacy
leaking and simply not useful for diff-ing.

The tool used to do the job is a personal project "Clean Filter for SVG"
and is accessible at https://github.com/Lin-Buo-Ren/Clean-Filter-for-SVG
, the actual code that do the XML processing: [1].  HTML Tidy must be run
with the contrib/svg.tidy config to restore the current markup style.

[1] https://github.com/Lin-Buo-Ren/Clean-Filter-for-SVG/blob/0ff326b/Clean%20Filter%20for%20SVG.bash#L351-L432

Signed-off-by: 林博仁(Buo-ren, Lin) <Buo.Ren.Lin@gmail.com>