Makefile: list sources when running flake8
The user might have virtualenvs and/or python installations (via
contrib/build-git-cola.sh) in their worktree, so invoking flake8 with
'.' is too wide a net and can be very slow.

We expect virtualenvs and tox environments to exist, per the
.gitignore rules, but the flake8 invocation would pick them up.

It's also unclear whether the `--exclude` flag replaces, or
augments the built-in exclusion list.  From the looks of it, the
flag full-on replaces the built-in patterns, which means that we
would also need to list `.git`, `__pycache__`, and other
possible directories in the exclusion list.

Maintaining an exclusion list in a world of virtualenvs is more
effort than using $(ALL_PYTHON_DIRS) in the target.

In light of this, it's better to be consistent with the rest of
the Makefile targets.  Use $(ALL_PYTHON_DIRS) and add
`contrib/win32` to the list of checked python directories so
that the pynsist preamble is picked up by pylint and flake8.

Related-to: #879
Reverts: af2636999f5d6960881d6d7fe064e188d0da482a
Signed-off-by: David Aguilar <davvid@gmail.com>