widgets: use the default text cursor
The block cursor, while being larger and easier to see, has usability
issues that detracts from the benefits of a larger cursor.

One important issue is that block cursors are traditionally used to
indicate a difference between "Insert" and "Overwrite" mode.  Cola does
not have this distinction, and thus using a block cursor conveys the
wrong message to users.  Furthermore, it can be harder to tell exactly
where text will be inserted when using a block cursor.

Revert to using the default text cursor.

Closes #889
Suggested-by: @et304383 via github.com
Helped-by: Charles <char101@ui.ac.id>
Helped-by: Ville Skyttä <ville.skytta@iki.fi>
Signed-off-by: David Aguilar <davvid@gmail.com>