app: workaround broken Qt Wayland widgets in gnome-shell
Qt emits a warning when XDG_SESSION_TYPE=wayland and
XDG_CURRENT_DESKTOP=GNOME because it is currently an unsupported
configuration.

Unset XDG_SESSION_TYPE=wayland when in gnome-shell to silence
the warning.  We will need to revert this commit in the future
when [QTBUG-68619](https://bugreports.qt.io/browse/QTBUG-68619)
makes Qt Wayland widgets usable in gnome-shell.

Signed-off-by: David Aguilar <davvid@gmail.com>