text: defer calls to setStyleSheet()
Older versions of Qt run into infinite loops if setStyleSheet() is
called from particular contexts, likely an event handler that triggered
a refresh (e.g. a focus change via the event filter).

This can be avoided by deferring the setStyleSheet() call so that it
happens the next time the event loop runs.

Signed-off-by: David Aguilar <davvid@gmail.com>