themes: remove the dockwidget margin
Save valuable screen space (especially on smaller laptop displays) by
eliminating the margins used in the flat themes.  This makes the margins
closer to the default theme.

The min-height setting was causing the dockwidget titles to get cut off
on macOS, and doesn't seem helpful or needed as things seem to render
better without it.

Signed-off-by: David Aguilar <davvid@gmail.com>