branch: rework the ref sorting implementation
Store the "ref_sort" as a simple `int` in the configuration to both
simplify the logic and to separate the configuration values from the
implementation details.

We now store a simple integer instead of the string used by git
to do the sorting.  Make the sort setting robust to values that
don't exist by internally modding the int to a known range.

Move the Cycle/Toggle behavior from the Branches widget into a
command object so that the mutation is not done directly in the
widget.

Lastly, keep the view updated via notification rather than explicitly
calling refresh() in various places.  This simplifies both the
branches widget and the main widget since it no longer needs to
be responsible for calling refresh().

Related-to: #930
Signed-off-by: David Aguilar <davvid@gmail.com>