cmds.LaunchDifftool: fix gnome-terminal support
Newer versions of gnome-terminal (and possibly older versions too)
require the command to run to be specified as additional arguments
instead of as a single string argument.

Add a `cola.terminalshellquote` configuration variable so that users
that need to tweak this behavior for custom terminals can do so.

Fix the usage of gnome-terminal to no longer shellquote arguments when
launching in-terminal commands such as "git difftool".

Fixes #935
Reported-by: Andrew Hemming <andrew.hemming@sky.uk>
Signed-off-by: David Aguilar <davvid@gmail.com>