hidpi: handle git config storing ints
The gitcfg reader will transform the EChoice strings into ints.  Make
this work by teaching the ComboBox class to transform values during
set_value().

This fixes the problem where values were not getting updated in the UI
on restart, due to the value being an int and not getting found in the
widget's data.

Signed-off-by: David Aguilar <davvid@gmail.com>