hidpi: disable overrides when environment variables are defined
Detect when either of  QT_SCALE_FACTOR or QT_AUTO_SCREEN_SCALE_FACTOR
are defined in "Auto" mode and do not override them when they have
already been set by the user.

This makes "Auto" more magical, but it will also make it follow the
user's explicit configuration, which is less surprising overall.

Related-to: #963
Signed-off-by: David Aguilar <davvid@gmail.com>