build: fixup #! lines for private helper scripts
Add distutils commands to handle fixing up the #! lines for our private
share/git-cola/bin scripts (specifically, git-xbase).

This ensures that the #! lines are updated for git-xbase when building
with alternate pythons.  Previously, this script would retain the
original `env python`, which may not be the same python as was used
during installation.

Related-to: #991
Signed-off-by: David Aguilar <davvid@gmail.com>