remote: operate on the model directly instead of using a copy
The fetch/push/pull dialog previously operated on a copy of the main
model due to some early architectural decisions that have since changed.

Early on, the model needed to contain attributes that mapped to
widgets directly so that they could be kept in sync through a
generic update callback that attempted to keep the widgets in
sync with the model.  This led to the decision to use a copy of
the model so that we can avoid needing to add extra attributes
for just these dialog.

This subsystem no longer exists, and our use of the model is
non-destructive, so we can eliminate all of that complexity.
Operating on a copy causes notifications to be lost and prevents the
main model from updating itself in response to operations, since
the operations end up editing the copy only.

Eliminate the model copy that is created for the remote dialogs
and operate on the model directly.  This ensures that updates to
the model are seen by the rest of the application.

Closes #996
Reported-by: akontsevich on github.com
Signed-off-by: David Aguilar <davvid@gmail.com>