recent: update recent repo list when opening repo
Currently, the recent repo list is only updated when the application is
exited, due to the list being stored in the settings, which are saved on
exit. As well as newly opened repos not showing on the recent list, it
means that if you start on repo A, open repo B, then open repo C, and
finally exit, repo B will not have been added to the recent list.

With this change, when a repo is opened it will be added to the recent
list immediately, and will therefore show in the recent repo list in the
menu. In addition to that, a signal was created for the changing of the
worktree, so that the bookmarks widget for the recent repo list can be
updated.

Closes #937
Signed-off-by: Tim Brown stimut@gmail.com